Fixed vendored deps not being retained
The dagger repo dependency is accessed with http_archive (unlike all others) which means it retrieves the .gitignore file from the repo and stores it in third_party during vendoring. That .gitignore file is used just as any other would be, and it so happens to remove the root BUILD file from the dagger repo (due to case insensitivity). When other machines download this repo, the missing BUILD file in the vendored repo causes the build to fail.